requirements.txt
# Core
numpy>=1.23.0
pandas>=1.5.0
scipy>=1.9.0
Pillow>=9.0.0
PyYAML>=6.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0                    # EfficientNet via timm

# Image Processing
albumentations>=1.3.0
opencv-python>=4.7.0

# WSI Processing (optional — requires OpenSlide C library)
# openslide-python>=1.2.0
# Install C library first: https://openslide.org/download/

# Evaluation
scikit-learn>=1.2.0
torchmetrics>=0.11.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Explainability
# pytorch-grad-cam>=1.4.0      # alternative Grad-CAM library

# Jupyter
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.15.0
ipywidgets>=8.0.0

# Progress
tqdm>=4.64.0

# Mixed precision
# torch.cuda.amp is built into PyTorch >= 1.6 (no extra install needed)
